+++ tests/testthat/test-get_bom_observations.R
@@ -1,13 +1,17 @@
 test_that("bom download works", {
+   expect_no_error({
    dl_location <- tempdir()
-  dl_loc <-
+  suppressMessages({
+   dl_loc <-
      get_bom_observations(ftp_url = "ftp://ftp.bom.gov.au/anon/gen/fwo/",
                           download_location = dl_location,
-                          access_warning = FALSE)
+                          access_warning = FALSE)})
+   })
 
-  dat <-
+  suppressWarnings(dat <-
      merge_axf_weather(File_compressed = dl_loc,
                     File_axf = "IDQ60910.99123.axf",
                     File_formatted = "NTamborine.csv",
-                    base_dir = getwd())
+                    base_dir = getwd()))
+  file.remove("NTamborine.csv")
 })